+++ changelog.md
@@ -4,7 +4,8 @@
 - Solved issue where after team member was invited into team folder, email with urging the recipient to create new account was sent
 - You are now allowed to set price for metered billing in 3 decimal places
 - Solved UI issue with empty notification popup
-- If you send file request for native user via email option, the push notification will be send to the user
+- If you send file request for native user via email option, the push notification will be sent to the user
+- Ability to test your websocket (Broadcasting) connection via Admin / Server / Broadcasting tab
 
 ## Version 2.1.3
 #### Release date: 10. May 2022